parametrized module no longer needed, passing all params to bm calls

+++ blond/beam/profile.py
@@ -449,7 +449,7 @@ class Profile(object):
         # Convert to uint32t for better performance
         self.n_macroparticles = self.n_macroparticles.astype(dtype, order='C')
 
-        worker.allreduce(self.n_macroparticles, dtype=dtype)
+        worker.allreduce(self.n_macroparticles)
 
         # Convert back to float64
         self.n_macroparticles = self.n_macroparticles.astype(
